feat: add early git url validation
